requirements.txt
streamlit==1.39.0
requests==2.32.3
nltk==3.9.1
python-dotenv==1.0.1
